+++ Emulator/Files/Folder.h
@@ -53,10 +53,10 @@ public:
 
     PETName<16> collectionName() override;
     u64 collectionCount() const override;
-    PETName<16> itemName(unsigned nr) const override;
-    u64 itemSize(unsigned nr) const override;
-    u8 readByte(unsigned nr, u64 pos) const override;
-    void copyItem(unsigned nr, u8 *buf, u64 len, u64 offset) const override;
+    PETName<16> itemName(isize nr) const override;
+    u64 itemSize(isize nr) const override;
+    u8 readByte(isize nr, u64 pos) const override;
+    void copyItem(isize nr, u8 *buf, u64 len, u64 offset) const override;
     
     //
     // Accessing
